requirements.txt
Flask==3.1.2 
gunicorn==20.1.0 
requests==2.32.5 
geopy==2.4.1 
python-dotenv==1.2.1 
secure-smtplib==0.1.1
email-validator==2.0.0